[Core] Rename obj_lifecycle_mgr to object_lifecycle_manager
The abbreviation was a workaround for the bazel build command exceeding
the OS command-line length limit. Recent bazel changes lifted that
limit, so the file and target names no longer need to be shortened.

Renames:
- plasma/obj_lifecycle_mgr.{cc,h} -> object_lifecycle_manager.{cc,h}
- plasma/tests/obj_lifecycle_mgr_test.cc -> object_lifecycle_manager_test.cc
- bazel target :obj_lifecycle_mgr -> :object_lifecycle_manager
