+++ src/ray/object_manager/plasma/BUILD.bazel
@@ -101,7 +101,7 @@ ray_cc_library(
     copts = PLASMA_COPTS,
     linkopts = PLASMA_LINKOPTS,
     deps = [
-        ":obj_lifecycle_mgr",
+        ":object_lifecycle_manager",
         ":plasma_allocator",
         ":plasma_object_store",
         ":plasma_store",
@@ -118,7 +118,7 @@ ray_cc_library(
     srcs = ["store.cc"],
     hdrs = ["store.h"],
     deps = [
-        ":obj_lifecycle_mgr",
+        ":object_lifecycle_manager",
         ":object_manager_plasma_common",
         ":plasma_allocator",
         ":plasma_connection_protocol",
@@ -148,7 +148,7 @@ ray_cc_library(
     srcs = ["get_request_queue.cc"],
     hdrs = ["get_request_queue.h"],
     deps = [
-        ":obj_lifecycle_mgr",
+        ":object_lifecycle_manager",
         ":plasma_connection_protocol",
         "//src/ray/asio:instrumented_io_context",
         "//src/ray/common:id",
@@ -156,9 +156,9 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "obj_lifecycle_mgr",
-    srcs = ["obj_lifecycle_mgr.cc"],
-    hdrs = ["obj_lifecycle_mgr.h"],
+    name = "object_lifecycle_manager",
+    srcs = ["object_lifecycle_manager.cc"],
+    hdrs = ["object_lifecycle_manager.h"],
     deps = [
         ":plasma_allocator",
         ":plasma_connection_protocol",
