+++ src/ray/object_manager/plasma/get_request_queue.h
@@ -20,7 +20,7 @@
 #include "ray/asio/instrumented_io_context.h"
 #include "ray/common/id.h"
 #include "ray/object_manager/plasma/connection.h"
-#include "ray/object_manager/plasma/obj_lifecycle_mgr.h"
+#include "ray/object_manager/plasma/object_lifecycle_manager.h"
 
 namespace plasma {
 struct GetRequest;
