+++ src/ray/object_manager/plasma/object_lifecycle_manager.cc
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "ray/object_manager/plasma/obj_lifecycle_mgr.h"
+#include "ray/object_manager/plasma/object_lifecycle_manager.h"
 
 #include <memory>
 #include <string>
