+++ src/ray/object_manager/plasma/tests/BUILD.bazel
@@ -37,14 +37,14 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "obj_lifecycle_mgr_test",
+    name = "object_lifecycle_manager_test",
     srcs = [
-        "obj_lifecycle_mgr_test.cc",
+        "object_lifecycle_manager_test.cc",
         "stats_collector_test.cc",
     ],
     tags = ["team:core"],
     deps = [
-        "//src/ray/object_manager/plasma:obj_lifecycle_mgr",
+        "//src/ray/object_manager/plasma:object_lifecycle_manager",
         "@com_google_absl//absl/random",
     ],
 )
